pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "prebuilt-custom-streaming"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "jupyter>=1.1.1",
    "langchain>=0.3.26",
    "langchain-openai>=0.3.27",
    "langgraph>=0.5.2",
    "langgraph-cli[inmem]>=0.3.4",
    "langgraph-supervisor>=0.0.27",
    "python-dotenv>=1.1.1",
]

[tool.setuptools.package-dir]
"" = "src"
